Load .glsl files in the data folder
This currently hard-codes the extension rather than asking the Mode for a list of supported extensions in the data folder. This would also need changes to the ChangeDetector, because this is only operating on filenames supposed to be in the sketch directory. (This could perhaps be converted to be using File objects?)